# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Sistem Pelaporan Prestasi Mahasiswa API
  description: |
    API untuk sistem pelaporan prestasi mahasiswa dengan fitur:
    - Authentication & Authorization (JWT)
    - Role-based Access Control (RBAC)
    - Achievement Management
    - Statistics & Reporting
    - User Management
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.prestasi.example.com/api/v1
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SubmitAchievementRequest:
      type: object
      required:
        - achievement_type
        - title
        - description
      properties:
        achievement_type:
          type: string
          enum: [akademik, non-akademik]
          example: "akademik"
        title:
          type: string
          example: "Juara 1 Programming Contest"
        description:
          type: string
          example: "Kompetisi programming tingkat nasional"
        details:
          type: object
          example:
            level: "nasional"
            category: "programming"
            date: "2024-01-15"
    AchievementResponse:
      type: object
      properties:
        message:
          type: string
          example: "achievement created successfully"
        data:
          type: object
          properties:
            achievement_id:
              type: string
              example: "507f1f77bcf86cd799439011"
            achievement_reference_id:
              type: string
              format: uuid
            student_id:
              type: string
              format: uuid
            achievement_type:
              type: string
              example: "akademik"
            title:
              type: string
              example: "Juara 1 Programming Contest"
            description:
              type: string
              example: "Kompetisi programming tingkat nasional"
            status:
              type: string
              enum: [draft, submitted, verified, rejected, deleted]
              example: "draft"
            created_at:
              type: string
              format: date-time

    # Statistics Models
...
paths:
...
  /achievements:
    post:
      tags:
        - Achievements
      summary: Submit prestasi (FR-003)
      description: Mahasiswa submit prestasi baru dengan status draft
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAchievementRequest'
      responses:
        '201':
          description: Achievement created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AchievementResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Insufficient permissions

    get:
      tags:
        - Achievements
      summary: Get my achievements
      description: Mahasiswa melihat prestasi sendiri
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, verified, rejected]
      responses:
        '200':
          description: List of achievements
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      achievements:
                        type: array
                        items:
                          $ref: '#/components/schemas/AchievementResponse'
                      pagination:
                        type: object
...
  /api/achievements/{reference_id}:
    delete:
      tags:
        - Achievements
      summary: Hapus prestasi (FR-005)
      description: Soft delete prestasi dengan status draft
      security:
        - BearerAuth: []
      parameters:
        - name: reference_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Achievement deleted successfully
        '400':
          description: Invalid request
        '404':
          description: Achievement not found
...